Separate home and foreign symbol in BinanceConfig
