+++ src/exchange/binance/mod.rs
@@ -4,7 +4,9 @@
 pub mod message;
 pub mod rest;
 pub mod ws;
+pub mod config;
 
 mod market;
 pub use market::*;
+pub use config::*;
